chore(deps): bump wgpu 23 -> 29, egui 0.30 -> 0.34
Supersedes dependabot #14. wgpu 23 -> 29 is a large jump with lots of
breaking API changes; egui-wgpu 0.34 pulls wgpu 29 so egui/egui-wgpu/
egui_extras also bump 0.30 -> 0.34.

Breaking-change fixes:

wgpu 29
- Instance::new: InstanceDescriptor no longer has Default; use
  InstanceDescriptor::new_without_display_handle() as the base.
- Adapter::request_adapter: now returns Result, not Option.
- Adapter::request_device: single descriptor argument; DeviceDescriptor
  gains `experimental_features` (ExperimentalFeatures::disabled()) and
  `trace` (Trace::Off) fields.
- PipelineLayoutDescriptor: bind_group_layouts is
  &[Option<&BindGroupLayout>]; push_constant_ranges removed, replaced
  by `immediate_size: u32`.
- RenderPipelineDescriptor: multiview -> multiview_mask
  (Option<NonZeroU32>).
- RenderPassDescriptor: new `multiview_mask` field.
- RenderPassColorAttachment: new `depth_slice: Option<u32>` field.
- ImageCopyTexture/ImageDataLayout renamed to
  TexelCopyTextureInfo/TexelCopyBufferLayout.
- Surface::get_current_texture: returns CurrentSurfaceTexture enum
  (Success/Suboptimal/Timeout/Occluded/Outdated/Lost/Validation), not
  Result.
- Device::push_error_scope returns an ErrorScopeGuard; pop via
  guard.pop() instead of device.pop_error_scope().

egui 0.34
- Rounding is a type alias for CornerRadius whose corner fields are
  u8, not f32. All `Rounding::same(x.0)` / `nw: x.0` sites updated to
  integer values or `as u8` casts.
- Margin fields are i8 (Margin::same takes i8).
- Painter::rect_stroke takes a fourth StrokeKind argument.
- Context::fonts closure receives &FontsView; `has_glyph` needs
  &mut self, so the bundled-fonts test uses fonts_mut.
- egui_wgpu::Renderer::new(device, format, RendererOptions) — single
  options struct replacing the depth/msaa/dither positional args.

+++ crates/juballer-deck/src/widget/builtin/dynamic.rs
@@ -237,8 +237,8 @@ fn render_node(ui: &mut egui::Ui, node: &ViewNode, cx: &RenderCx<'_>) {
             let rad = rounding.unwrap_or(0.0);
             egui::Frame::none()
                 .fill(fill)
-                .rounding(egui::Rounding::same(rad))
-                .inner_margin(egui::Margin::same(4.0))
+                .rounding(egui::Rounding::same((rad) as u8))
+                .inner_margin(egui::Margin::same(4))
                 .show(ui, |ui| render_node(ui, child, cx));
         }
         ViewNode::Progress {
@@ -433,7 +433,7 @@ fn render_plot(
     let painter = ui.painter_at(rect);
     painter.rect_filled(
         rect,
-        egui::Rounding::same(2.0),
+        egui::Rounding::same(2),
         egui::Color32::from_rgba_unmultiplied(0x1e, 0x1e, 0x2e, 128),
     );
     if values.len() >= 2 {
